pyproject.toml
[project]
name = "stock-analysis-agent"
version = "0.1.0"
description = "Multi-agent stock analysis system for Indian markets (NSE/BSE) powered by Google ADK and Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # --- Agent Framework ---
    "google-adk>=0.5.0",
    "google-genai>=1.0.0",

    # --- Stock Market Data ---
    "yfinance>=0.2.40",

    # --- Web Search ---
    "tavily-python>=0.5.0",
    "ddgs>=6.0.0",                       # free fallback (formerly duckduckgo-search)

    # --- PDF Parsing ---
    "pdfplumber>=0.11.0",
    "PyMuPDF>=1.24.0",

    # --- RAG Pipeline ---
    "chromadb>=0.5.0",
    "langchain-text-splitters>=0.2.0",   # smart chunking only

    # --- Data Processing ---
    "pandas>=2.2.0",

    # --- UI ---
    "streamlit>=1.37.0",

    # --- Config & Utilities ---
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",                      # pretty terminal output
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]
